extcap: Check for valid PID before killing child process
If an extcap tool fails to start or the tool exits unexpectedly
Wireshark will try to kill PID -1. This has very unexpected results
on Linux, like bringing down the whole window manager.

Make sure it's a valid PID before killing the extcap child process.

Bug: 16362
Fixes: v3.3.0rc0-461-g8efde39805 ("extcap: terminate the child process using kill.")
Reviewed-on: https://code.wireshark.org/review/35998
Petri-Dish: Pascal Quantin <pascal@wireshark.org>
Petri-Dish: Guy Harris <guy@alum.mit.edu>
Tested-by: Petri Dish Buildbot

+++ extcap.c
@@ -1220,7 +1220,10 @@ void extcap_if_cleanup(capture_options *capture_opts, gchar **errormsg)
         /* Send termination signal to child. On Linux and OSX the child will not notice that the
          * pipe has been closed before writing to the pipe.
          */
-        kill(interface_opts->extcap_pid , SIGTERM);
+        if (interface_opts->extcap_pid != WS_INVALID_PID)
+        {
+            kill(interface_opts->extcap_pid, SIGTERM);
+        }
 #endif
         /* Maybe the client closed and removed fifo, but ws should check if
          * pid should be closed */
